fix: resolve stale content issue in AI metadata generation by sending blogText
