+++ server/src/types/index.ts
@@ -11,12 +11,6 @@ export interface MailOptions {
   html?: string;
 }
 
-export interface EditorJsBlock {
-  id: string;
-  type: string;
-  data: Record<string, any>;
-}
-
 export interface BlogMetadata {
   title: string;
   summary: string;
